backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.2
SQLAlchemy>=2.0
aiomysql>=0.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt<4 # passlib 1.7.4 兼容区间
httpx>=0.27

# nanobot（作为后端进程内依赖运行）
nanobot-ai>=0.1.5.post1
pandas>=2.0
numpy>=1.24
tabulate>=0.9
statsmodels>=0.14

